Introduce understandable claim names
Using our internal identifiers as external identifiers is a bit of a
pain, because all of our users will need to keep around a mapping of
incomprehensible UUIDs to understandable names.

This commit introduces claim names, included in the same config file
as the permissions, which are used for external communication.  I've
tried to reduce the number of places where we need to be aware of the
difference between external names and internal identifiers.