Implement an endpoint to set multiple attributes
We can use this in the RemoteUserInfo.update_profile! method in
account-manager to set both the `email` and `email_verified`
attributes in one fell swoop, which will reduce the total number of
network round-trips from 4 (2 requests to attribute-service, each of
which makes a request to account-manager to check the token) to 2.