Add a migration to correct the invalid claims data
All claims created with the bulk endpoint will be strings in need of
parsing.  Claims updated through the non-bulk endpoint (transition
checker state set directly by the checker, rather than set by the
account manager as a result of registration) will be stored correctly,
and so will throw an error when we try to parse them.