Update ruby to 2.7.3 and bundler to 2.2.17
This is the solution to [this deployment failure][1]. The build was
complaining about ruby and bundler versions:

```
**WARNING** A newer version of ruby is available in this buildpack.
Please adjust your app to use version 2.7.3 instead of version 2.7.2 as
soon as possible. Old versions of ruby are only provided to assist in
migrating to newer versions.
```

```
**WARNING** Your Gemfile.lock was bundled with bundler 2.1.4, which is
incompatible with the current bundler version (2.2.16).
```

[1]: https://cd.gds-reliability.engineering/teams/govuk-tools/pipelines/govuk-attribute-service-prototype/jobs/deploy-app-staging/builds/141

Co-authored-by: Alan Gabbianelli <Alan.Gabbianelli@digital.cabinet-office.gov.uk>
Co-authored-by: Huw Diprose <Huw.Diprose@digital.cabinet-office.gov.uk>