Add a show_email_confirmation_reminder attribute
On the "your information" page in the account manager, we show a box
reminding the user to confirm their email address.

This box shows up if the user is not confirmed (in which case the
`email_verified` attribute will be false) OR if the user is
confirmed (in which case the `email_verified` attribute will be true)
BUT they have provided a new email address which they haven't verified
yet.

So to move the "your information" page out of the account manager and
onto GOV.UK, we need a new attribute to handle this second case: the
user has confirmed the email address they registered with, but they
have a new one pending.