Remove some more unused bits of Rails
After upgrading to Rails 6.1, it seems like one of these pulls in
ActiveStorage when run with RAILS_ENV=production (but not
RAILS_ENV=development or test), which is throwing an error because we
don't have a config file for it.

I've commented out and removed unused bits of Rails, to match the
account-api set-up, which fixes the issue.